Merged PR 2121: CompositionBrush Overview overhaul
Initial pull request to get early feedback on CompositionBrush Overview doc. Remaining TODOs
1. Update images for each of the brush types
2. Clean up remaining links

I am also planning to add a bunch of new docs for some of these brushes and howto's which might impact where in the links this doc lives, but that is planned after this initial update goes out.