Merged PR 5432: Merge giacosta/xim-documentation-edits to master
This is a large edit of the structure of XIM documentation. In order to make it be more aligned to how other sections of documentation are formatted, I added TOCs to the top of most of the documents. My editor has a Markdown Validator, so invalid Markdown code was removed from changed files. More headers were added to further split up content. Changes were aligned to C# version. Comments are distributed within using the <!-- prefix where discussion or TODOs can be done. XR compliancy section was added to overview. "Overview" term removed from md. file name so as not to be the only one that has it. Reordering of sections was done in main Using XIM and using XIM CS file to improve discoverability, perusing capability, and readability.