Update navigationview.md
There is a bug in the AppTitle code that does not handle FullScreenMode. I have updated the sample code to account for this. It might even be a bug that the LayoutMetricsChanged does not raise when FullScreenMode changes, and TitleBar still reserves space for the BackButton even though it is no longer visible. In either case, the code sample accounts for both scenarios.