Update auto-build-package-uwp-apps.md
Recently I used VSTS for setting up CI for an UWP app, according to the docs an appx file has to be used for setting up automated testing on the build but from my recent experience an .appxrecipe is required. Also in the "other console options" the parameter /framework:FrameworkUap10 had to be added to make the tests run. If you'd like build logs from this configuration, I'd be happy to provide.