Merged PR 6149: Overview documentation for using the c# gamechat2 projection
This PR adds the overview documentation for using Game Chat2 with C# projects.

Also, I am not sure if now is the right time to actually check in the PR since the feature hasn't actually shipped yet.

The work item for this is [8536087](https://microsoft.visualstudio.com/Xbox/_workitems/edit/14319577), but is in a different project, so doesn't seem to be showing up in the Work Items section.