Merged PR 7504: CR - Point of Service Documentation - Camera Barcode Scanner
Organizes the PointOfService Documentation landing page with topic categories and introduces basic barcode scanning as well as camera barcode and sample barcodes supported by the in-box decoder.

I am most interested in your feedback on the following three files

| file | comment |
|----|----|
| pos-barcodescanner-get-started.md | This document is going to be broken down into more specific topics in the near future, so I care more about accuracy than adding code snip-its or completeness for now. |
| pos-camerabarcode-getstarted.md | pay closer attention to this one |
| pos-camerabarcode-hosting-preview.md | pay closer attention to this one |

Welcome feedback on all, but these three are the most important in this CR.

To view the MD preview, CTRL + Click the file in the PR and select Preview.

- Created initial pos-barcodescanner.md
- Created initial pos-camerabarcode.md
- Created initial pos-camerabarcode.md
- Added list of supported symbologies
- Added sample barcode images
- Updated landing pages, camera barcode topics and added sample barcodes
- Deleted desktop.ini - Added by mistake
- Spell and grammar check complete
- Minor touch ups to add consistency between topics