Merged PR 9175: Visual Studio Version handling.
Addressing latest Visual Studio concerns where instances of Visual Studio 2015 are mentioned.

A large portion of our documents were written when Visual Studio 2015 was the latest release of Visual Studio or it was still in popular use. There are many documents that speak to Visual Studio 2015 that were not touched but this Pull Request addresses many instances where it seems appropriate to note that the LATEST iteration of Visual Studio should be used. Currently Visual Studio 2017. The points of most concern are those instances where we mention Visual Studio as a development requirement.