Merged PR 9287: adding blurb to the get the xdk doc that tells where to get the xdk
adding blurb to the get the xdk doc that tells where to get the xdk

Need to fill in a few blanks if applicable.
Can a company apply to be a managed partner?
Are there other groups of people with XDK access?
Do we want to describe downloading and installing the XDK instead of directing developers to their account manager?