Fixing typos in gaze-interaction sample
Just used this sample at a recruiting event. When copy/pasting into Visual Studio I found two typos. One was a "..." in the first code snippet. Not sure of the purpose of this, it shows as an error in the code editor so I removed it. The second was "timersStarted" which had two s letters in it.