Merged PR 11998: Managed partner pointed this out as an error. Please verify?
When following up on a forum post discussion, a managed partner pointed out this line as being in error with this suggested fix.

I'm not sure who the feature PM is for this... can one of you please verify (or find the proper SME who can verify) if this suggested change is accurate or not? If it's not correct, please reject this PR; if yes, approve. Thanks!