Merged PR 12963: Migrated change tracking docs
Sorry for the duplicate - the first PR didn't have the images in it.

Migrating an existing article from my MSDN blog to the proper docs. This article covers how to use the StorageLibraryChangeTracker to change track a library using a UWP app

There will be a few commits like this coming in the next few weeks, as I'm trying to get all the useful information off of the blog and onto the docs site before the blogs are taken down.