Stated dimensions
I added 1 line to explicitly state that the listed dimensions are width x height.

The included images are confusing and do not sufficiently communicate dimensions.  For example, the first image states 100 x 600 and shows a landscape-oriented screen. That implies that the dimensions are height x width (as the image is wider than it is high), however a bright blue box inside the image is portrait-oriented. The listed dimensions in fact refer only to that section. That's too many hurdles for the reader to jump. Explicitly stating width x height avoids possible confusion.