Fixed confusing document
Revise the style of the keypress representation to match that of the mainstream standard.

The main reason is that "alt+shift+plus" is confusing. This internal Windows Terminal's notation leads the readers to believe that they must press the NumPad Plus key while holding down Alt and Shift. **Not so!** The "alt+shift+plus" notation actually means pressing the Equal Sign (=) while holding down Alt and Shift keys.