dholloweCameraScannersVsCameraScanners
Some dedicated barcode scanners use a camera. This article refers specifically to the symbologies supported by the software decoder that is used with standard lens cameras. Additionally symbologies can be supported via camera barcode scanners, but point out the fact that this is specific to scenario where a web cam is being used as if it were a barcode scanner.
It's also possible for additional symbologies to be introduced in newer versions of Windows 10. We should go ahead and tell the customer which version of the OS introduced supported for each individual symbology, instead of having customers assume that all of the symbologies are supported since 17134 when the barcode decoder was originally added.