Change Protocol Activation Parameter from Optional
I work at Microsoft (on the Windows Media Apps team, on Photos) and recently encountered an issue when trying to launch a video for editing with this URI scheme. I was only specifying the InputToken parameter here, since the others were indicated as optional, and my video would never open, I just get a black screen. Feel free to try it with the Windows Store build of Photos and you'll see this behavior. I dug around in the code and it turns out that you do need to specify this parameter for the videoedit activation to work. As such, it should be changed from an optional parameter to a required one.