Update broken link
"Get set up" link is broken. However, there is a "Get set up" section in the /windows/uwp/get-started/ page. I assume the link pointed to a stand-alone page but then moved into a section of the "Get started" page. Thus, this update it changes the link along with the link text to describe the location properly.