Add required version for OMP2
Oh-my-posh has updated to version 3, so adding a note to designate version 2 on this uninstall until we can return to this tutorial and update to version 3. In the meantime, there are other tutorials available on the web that walkthrough how to install OMP v3 and set it up with git status, etc.