Updating get started instructions  (#1774)
* deleting instructions for explicitly setting the .NET SDK version

This step is no longer required and the FrameworkReference isn't supported anymore.

* removing note about update

The instructions are vague, it didn't actually work, and we don't provide explicit instructions on how to validate that it did work.

We can add detailed instructions to replace this later. Will send a note to Andrew/Scott to help validate how an automatic update would work.