Corrected opening sentence in Unplated assets
The "Unplated assets" paragraph started out with "By default, Windows uses a target-based asset on top of a colored backplate by default".  By starting a sentence with "By default" it is unnecessary to end the same sentence with "by default".  This change removes the second iteration of "by default" in favor of the first.