Update deployment docs (#2338)
* Update deploy-packaged-apps.md

* Update deploy-unpackaged-apps.md

* Editing

It's best for link text to identify the destination. The link text doesn't have to be an *exact* match for the title; just close enough to make it clear and unambiguous where the link will take the customer. If we hyperlink things like just "single-project MSIX", then that makes it look like it goes to some glossary definition. In any case, customers won't know where it goes, because the link text is not the same as the title of the destination.

If link text doesn't tell a customer what the title of the destination is, then they won't know whether they've been there before, so they'll liklely click it to find out, find that they've been there before (once they see the actual title), and then just have to go back to what they were reading. There's been lots of feedback that recent docs are a confusing (overwhelming) graph of links that cycle around, and are just generally unpredictable. I strongly suspect that using good link text will alleviate that.

information -> info

I changed "For C# apps," back to "**C#**." Bold little qualifiers like that are much easier to spot at a glance.

* Edits

Added clarity around MISX-packaged, and then non-MSIX-packaged (which includes sparse-packaged and unpackaged). Using the term "unpackaged" to include "sparse-packaged" is a problem, and this fixes that.

I changed "For C# apps," back to "**C#**." Bold little qualifiers like that are much easier to spot at a glance (and they're an established pattern).

* Updating ownership

* Updating ownership

Co-authored-by: Steven White <31261191+stevewhims@users.noreply.github.com>