[BrokenLinksH2] missing page
As you can see, I removed the bottom bullet point in the list of reports.  After extensive searching I could not find a report on ad campaigns themselves.  The closest I was able to find is the advertising performance report in the bullet above.

**Global effort to fix broken links**

The Content & Learning team is fixing broken links on docs.microsoft.com for the rest of H2. This effort will eliminate potential accessibility, security, and usability issues. This PR includes only link fixes and does not change other content. Please review within five business days and merge, or comment in the PR with any changes you'd like to see. Thanks!