Fix C# code compile erros
This fix doesn't make this sample working. It just makes the C# code be able to be compiled.

---

**Other issues**

- There is no C# code in [Step 5](https://docs.microsoft.com/en-us/windows/apps/windows-app-sdk/notifications/app-notifications/app-notifications-quickstart?tabs=cs#step-5-process-a-user-selecting-a-notification).
- There is no C# code in [Step 6](https://docs.microsoft.com/en-us/windows/apps/windows-app-sdk/notifications/app-notifications/app-notifications-quickstart?tabs=cs#step-6-remove-notifications).
- In *Step 3* there is reference to `ToastWithTextBox.ScenarioId` which is not implemented anywhere.
- Calling `AppNotificationManager.Default.Register();` throws *Error HRESULT E_FAIL has been returned from a call to a COM component* error.

---

For me *Getting started* example could loook something like:

```cs
public static bool SendNotificationToast(string title, string message)
{
    var xmlPayload = new string($@"
        <toast>    
            <visual>    
                <binding template=""ToastGeneric"">    
                    <text>{title}</text>
                    <text>{message}</text>    
                </binding>
            </visual>  
        </toast>");

    var toast = new AppNotification(xmlPayload);
    AppNotificationManager.Default.Show(toast);
    return toast.Id != 0;
}

It's good starting point for the developer wanting to notify the user about something, where it's not important what the user will do with the notification. In such case there is no need for updating the `Package.appxmanifest` file, nor to build `NotificationManager`. Single function with few lines of code is enough for that :-) 
```