Update app-submissions.md (#2837)
* Update app-submissions.md

On Aug 1, publishing and certification notifications for apps and games developers will be migrated to Action Center, a new communication hub within Partner Center. With the launch, we are also introducing two changes, notifications will be delivered to developer account owner instead of seller account contact and notifications will be accessible through Action Center in addition to emails. 

We would like to get this change published on Aug 2nd, PDT.

* Update app-submissions.md

Converted link to relative.

Co-authored-by: John Kennedy <jken@microsoft.com>