More suggested changes to improve this page
Getting started should be simple enough that the reader will be successful with the minimal amount of steps. This Getting started is missing a summary and what the end goal is. Limit the number of choices the reader has to do. Either start with a new project (recommended) or how to modify an existing project, not both. Giving the reader choices only confuses the reader. If the reader is unsuccessful, then the reader will abandon this technology for being too hard. So now you have an uphill battle to win the reader back. Make them win the first time out.