Addressing review comments
I think I've address all of the comments/suggestions. We may need to iterate a bit on the "processor architecture" section I added.
One question... what would be the best format to choose for the code block on lines 211-218? I picke YAML because it looks good with that format. Would INI be better? The markdown linter wants me to pick something.