Rephrased question in hub/dev-drive/index.md
Question `Can I only have a single Dev Drive on my machine?` can be interpreted in a few ways by the reader.

1. The question can be interpreted as following: Can I have a PC configuration which has just one dev-drive (no less no more)? Will this configuration work?

That's how I interpreted it initially. And I was quite surprised when the first word of an answer was `No.` which meant that I have to have either 0 dev-drives or ≥2 dev-drives. And after that I've read an explanation that implied you clearly meant other interpretation of the question.

2. The second interpretation is yours. (And I actually still don't quite understand as non-native how the question can be interpreted that way)

So I rephrased the question and the answer for clarification.