Main > Live (#1894)
* Fixing small formatting and writing errors. (#1863)

* Fix broken code block and bookmark link (#1866)

* Fix broken code block

* Update broken wslconfig bookmark

* Update install-on-server.md (#1871)

Fix indent and numbering for better readability.

* Fix punctuation errors (#1876)

* small improvements in troubleshooting.md (#1877)

Some small formatting improvements.
Also added the command to restart the hns service.

* Fix a few typos (#1879)

Co-authored-by: Matt Wojciakowski <mattwoj@microsoft.com>

* Updated guide to fix an issue because of a wrong command (#1884)

* Update wsl-containers.md

Running Dev Containers: **Open Folder in Container...** when using this guide is not / no longer correct, as most readers will already have opened the folder while connected to WSL based on the steps provided above. We need to use **Dev Containers: Reopen in Container** or it will fail to create the .devcontainer folder and devcontainer.json and give a (unfortunately not very clear) error message.


See also: https://code.visualstudio.com/docs/devcontainers/containers#_open-a-wsl-2-folder-in-a-container-on-windows
   
- Use the Dev Containers: Reopen in Container command from a folder already opened using the WSL extension.
- Select Dev Containers: Open Folder in Container... from the Command Palette (F1) and choose a WSL folder using the local \\wsl$ share (from the Windows side).

* Update wsl-containers.md

---------

Co-authored-by: Matt Wojciakowski <mattwoj@microsoft.com>

* Update faq.yml -- tiny grammar change (#1885)

* Update faq.yml -- tiny grammar change 

'changes to locale to' -> 'changes the locale to'

* Update basic-commands.md

'If you running WSL' -> 'If you run WSL'

* Update basic-commands.md

* rephrase a long sentence
* swap order of two sentences to make example of command closer to the description of the command

example of command: `wsl --set-default-version 2`

* Update basic-commands.md

Replace 

wsl -u <Username>`, `wsl --user <Username>

with

wsl --user <Username>

`wsl -u <Username>` and `wsl --user <Username>` are equivalent. Both commands are correct for the section of the documentation, but the display of the commands in the documentation was incorrect due to the stray ticks and comma.

* Update gui-apps.md to simplify commits. (#1880)

* Update gui-apps.md

If we are already referencing apt, why not just use apt to install google chrome?

* Update gui-apps.md

forgot --fix-missing

* Add a note to explain the command flags and path

---------

Co-authored-by: Matt Wojciakowski <mattwoj@microsoft.com>

* Fix anchor warning

---------

Co-authored-by: Stefan Nikolaj <51296436+snikolaj@users.noreply.github.com>
Co-authored-by: Anton Kesy <antonkesy@gmail.com>
Co-authored-by: Ishan Jain <github@ishan.jain.se>
Co-authored-by: Krishna Vivek Vitta <84629581+krishnacx@users.noreply.github.com>
Co-authored-by: icyerasor <andreas@feldschmid.com>
Co-authored-by: Blake-Madden <66873089+Blake-Madden@users.noreply.github.com>
Co-authored-by: vmsteiner <107506233+vmsteiner@users.noreply.github.com>
Co-authored-by: Andrew Zipperer <47086307+zipperer@users.noreply.github.com>
Co-authored-by: Wes Bryie <wes@wesbryie.tech>